Start moving InMemoryFileIndex over to HadoopFSUtils
